requirements.txt
asgiref==3.2.7
Django==3.0.6
django-phone-field==1.8.0
django-taggit==1.3.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.13.0
pytz==2020.1
sqlparse==0.3.1
